Move docket module into notice_comment.
